% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{compilation}
\alias{compilation}
\title{Grouping data into form suitable for heatmap}
\usage{
compilation(data, y, Ident, z, index, t.sort = NA, N, not.q = F, t.func,
  q.probs, grp.func = mean, q.wgt = NA)
}
\arguments{
\item{data}{A data.frame, or data.table.}

\item{y}{The factor/quantile column (y-axis on heatmap).}

\item{Ident}{The individual identifier, creating unique key with time.}

\item{z}{The dependent variable plotted (variation in heatmap)}

\item{index}{The time/index column (x-axis in heatmap)}

\item{t.sort}{If instrument is time-varying, the time period where sorting
is made.}

\item{N}{Number of quantiles on y-axis.}

\item{not.q}{Default \code{FALSE}, set to \code{TRUE} if y is character,
at which point the column will be set to a factor later.}

\item{t.func}{Output from \code{\link{timeConv}}.}

\item{q.probs}{Optional, set quantiles at which instrument bins are cut.}

\item{grp.func}{Aggregation function, default is \code{mean}.}

\item{q.wgt}{Weighted quantiles using column from the dataset.}
}
\value{
Collapsed dataset of same class as input, with columns "time", "y",
"quantile."
}
\description{
\code{compilation} groups primitive data into a heatmap-suitable form,
i.e. aggregation along the time/index and factor/quantile axes.
}
\details{
This function first searches for the time/index column and retains it.
Then it drop duplicates of ID, either binning the IDs into quantiles
(analogous to Stata's xtile function) or preserving a factor variable
to be converted later.

This part of the code is run in Stata/Python for their respective ports.
One difference is that the aggregation function allowed in R can be much
more creative than what are admissible in Stata.
}
\keyword{internal}

