% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{parseHeatform}
\alias{parseHeatform}
\title{Formula parsing function for heatmap.}
\usage{
parseHeatform(relation)
}
\arguments{
\item{relation}{Formula of the form Y ~ CrS(X,ID,[w]):i[(t)]. Y is the
dependent variable to be plotted in colour. The arguments of CrS are,
left to right, the instrument/factor variable, the cross-sectional identifier
which should form a unique key along with time, and a weighing variable
for each observation. i is the index/time column, and if (t) is specified i
is assumed to be time varying, with data being sorted on values of i at
time t.}
}
\value{
A list of parsed arguments for entry in \code{\link{heatmapEco}}.
}
\description{
\code{heatmap} needs to interpret a particular kind of formula to understand
which columns contain data to be added on the axes, and so forth.
\code{parseHeatform} handles this.
}
\details{
Internal function. When inputted, "CrS" could be replaced by any other name,
so long as the formula looks like it has a named function within.
}
\examples{
\dontrun{
parseHeatform(Y ~ CrS(x,ID):time) # Sorting on time-invariant instrument
# Sorting on time varying instrument at time 2001
parseHeatform(Y ~ CrS(x,ID):time(2001))
# The following will not parse correctly
parseHeatform(Y ~ (x,ID):time)
parseHeatform(Y ~ CrS(x,ID):.)
}
}
\keyword{internal}

