% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{setupFill}
\alias{setupFill}
\title{Setting up fill aesthetic of heatmap}
\usage{
setupFill(fill.col, outliers, zlab, count = F, custom)
}
\arguments{
\item{fill.col}{Input column of time values.}

\item{outliers}{If set, will add two low-luminosity colours to both ends of
the gradient, intended to help perceive observations exceeding the
1.5*IQR heuristic.}

\item{zlab}{Label for the fill variable to be placed above the legend.}

\item{count}{Vector of character/integers expressing time in the same
format specified by \code{\link{timeConv}}.}

\item{custom}{Percentiles at which the principal palette colours are placed.}
}
\value{
List of ggplot2 formulae to be parsed by \code{\link{heatmapBuild}}.
}
\description{
\code{setupFill} adds bells and whistles to the fill aesthetic (in ggplot
terms).
}
\details{
Heatmap accepts two colour gradient; a divergent red-blue palette
and a sequential red-yellow-pale blue palette. By default, the blue palette
has unequal variation biased to lower quantiles in order to accentuate
"Ashenfelter dips". Apart from the default palettes, other options are
adjustable.
}
\keyword{internal}

