% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{setupX}
\alias{setupX}
\title{Setting up X axis of heatmap}
\usage{
setupX(t.col, split, xtitle = "Time", period, time.axis, pol.break)
}
\arguments{
\item{t.col}{Input column of time values.}

\item{split}{If set, determines the number of gaps between each X-axis
label. Default spacings exist for each time type specified by
\code{\link{timeConv}}.}

\item{xtitle}{Label for the variable corresponding to the X-axis.}

\item{period}{Same as argument \code{period} from \code{\link{timeConv}}.}

\item{pol.break}{Vector of character/integers expressing time in the same
format specified by \code{\link{timeConv}}.}
}
\value{
List of ggplot2 formulae to be parsed by \code{\link{heatmapBuild}}.
}
\description{
\code{setupX} adds bells and whistles to aspects of the X axis.
}
\details{
Some of the actions done include spacing out labels; adjusting bounds given
the actual minima and maxima of the time variable; and plotting vertical
lines that show when policy changes occur.
}
\keyword{internal}

