% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{setupY}
\alias{setupY}
\title{Setting up Y axis of heatmap}
\usage{
setupY(y.col, split, ytitle = "Instrument", factor.ax = F)
}
\arguments{
\item{y.col}{Input column of categorical/quantile values.}

\item{split}{If set, determines the number of gaps between each Y-axis
label. The default spacing is to have 5 evenly-spaced labels.}

\item{ytitle}{Label for the variable corresponding to the Y-axis.}

\item{factor.ax}{If \code{TRUE}, interprets column as a factor.}
}
\value{
List of ggplot2 formulae to be parsed by \code{\link{heatmapBuild}}.
}
\description{
\code{setupY} adds bells and whistles to aspects of the X axis.
}
\details{
Actions currently possible include spacing out of labels and indication of
whether the Y-axis are factors (in which case every row is labelled).
}
\keyword{internal}

