% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{timeConv}
\alias{timeConv}
\title{Parse time variable to machine-readable formats.}
\usage{
timeConv(period, ft = "")
}
\arguments{
\item{period}{A string indicating the time interval. One of "yearmon",
"year" or "yearqtr."}

\item{ft}{R date format string as in as.Date.}
}
\value{
Function that converts string of format ft into an internal numeric
representing period.
}
\description{
\code{timeConv} is a closure allowing flexible parsing of a time column
into an appropriate format for the heatmap x-axis.
}
\details{
To increase ease of use, the \code{heatmap} package can parse
date and time strings, so long as the format of the string is specified.
Furthermore, the time interval represented by each block in the heatmap
can be set using \code{timeConv}. The function transforms the string
into a date class using native R functions, then converts it into integers
so the plot can be built properly.
}
\examples{
timeConv("yearmon")(c("2008-01-01", "1934-12-15"))
\dontrun{
timeConv("yearmon")("1993") # Have to put month there as well, even if dummy
timeConv("yearmon", "\%d\%b\%Y")("jan112011") # wrong formatting order
}
}

