% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapBuild}
\alias{heatmapBuild}
\title{Construct and export heatmap}
\usage{
heatmapBuild(data, xtick, xl, int, ytick, ylab, factor.ax, fill, zlab, save)
}
\arguments{
\item{data}{The collapsed dataset.}

\item{xtick}{Placement of ticks and labels on the X-axis, set with setupX.}

\item{xl}{X-axis boundaries, set with setupX.}

\item{int}{Shaded vertical lines (x-intercepts), set with setupX.}

\item{ytick}{Placement of ticks and labels on the Y-axis, set with setupY.}

\item{factor.ax}{If \code{TRUE}, interprets variable column as a factor for
aggregation and options.}

\item{fill}{Colour gradient derived from palette, set with setupFill.}

\item{zlab}{Label of dependent variable for legend, set with setupFill.}

\item{save}{File to which graph is exported using ggsave.}
}
\description{
\code{heatmapBuild} compiles the heatmap given settings.
}
\details{
This is an internal function. Its arguments are the collapsed data and
elements of the list outputs from \code{\link{setupX}} or
\code{\link{setupFill}}
}
\keyword{internal}

